change from double to float

+++ src/bindings.cpp
@@ -8,17 +8,17 @@ namespace py = pybind11;
 PYBIND11_MODULE(SPO2Algorithm, m) {
   m.def(
     "get_oxygen_level",
-    [](py::array_t<double, py::array::c_style | py::array::forcecast> ppg_ir,
-       py::array_t<double, py::array::c_style | py::array::forcecast> ppg_red,
+    [](py::array_t<float, py::array::c_style | py::array::forcecast> ppg_ir,
+       py::array_t<float, py::array::c_style | py::array::forcecast> ppg_red,
        int sampling_rate,
-       double callib_coef1,
-       double callib_coef2,
-       double callib_coef3)
+       float callib_coef1,
+       float callib_coef2,
+       float callib_coef3)
     {
-        double oxygen_level = 0.0;
+        float oxygen_level = 0.0;
         get_oxygen_level(
-            const_cast<double*>(ppg_ir.data()),
-            const_cast<double*>(ppg_red.data()),
+            const_cast<float*>(ppg_ir.data()),
+            const_cast<float*>(ppg_red.data()),
             static_cast<int>(ppg_ir.size()),
             sampling_rate,
             callib_coef1,
